+++ src/users/hospital/hospital.controller.ts
@@ -7,6 +7,7 @@ import { Claims, Role } from '../../auth/models/claims.interface';
 import { Auth } from '../../auth/decorators/auth.decorator';
 import { UsersService } from '../users.service';
 import { Roles } from '../../auth/decorators/roles.decorator';
+import { DoctorService } from '../../doctor/doctor.service';
 
 @Auth
 @Controller('user')
@@ -14,7 +15,8 @@ export class HospitalController {
 
     constructor(
         private hospitalService: HospitalService,
-        private userService: UsersService
+        private userService: UsersService,
+        private doctorService: DoctorService
     ){}
 
     @Get('hospital-treatments')
@@ -30,42 +32,53 @@ export class HospitalController {
     @Roles(Role.Doctor)
     async getSomeTreatments(
         @Param('userId') id: string, 
-        @Query('page') page: 1
+        @Query('page') page: 1,
+        @User() claims: Claims
     // ): Promise<{ hospitals: Hospital[]; count: number }> {
     ): Promise<Hospital[]> {
-        await this.userService.assertExists(id);
-        const [hospitals, count] = await this.hospitalService.findHospitalTreatments(id, page);
-        // return { hospitals, count };
-        return hospitals;
+        if (await this.doctorService.hasAccess(id, claims)) {
+            await this.userService.assertExists(id);
+            const [hospitals, count] = await this.hospitalService.findHospitalTreatments(id, page);
+            // return { hospitals, count };
+            return hospitals;
+        }
     }
 
     @Post(':userId/hospital-treatments')
     @Roles(Role.Doctor)
     async addHospitalTreatments( 
         @Param('userId') id: string,
-        @Body() hospital: HospitalBindings
-    ): Promise<Hospital>{      
-        await this.userService.assertExists(id);
-        return await this.hospitalService.addHospitalTreatment(hospital, id);
+        @Body() hospital: HospitalBindings,
+        @User() claims: Claims
+    ): Promise<Hospital>{
+        if (await this.doctorService.hasAccess(id, claims)) {
+            await this.userService.assertExists(id);
+            return await this.hospitalService.addHospitalTreatment(hospital, id);
+        }
     }
 
     @Put('hospitals/:hospitalTreatmentId')
     @Roles(Role.Doctor)
     async editHospitalTreatments(
         @Param('hospitalTreatmentId') treatmentId: string,
-        @Body() hospital: HospitalBindings
+        @Body() hospital: HospitalBindings,
+        @User() claims: Claims
     ){
-        return await this.hospitalService.editHospitalTreatments(treatmentId, hospital);
+        const treatment = await this.hospitalService.getUserId(treatmentId);
+        if (await this.doctorService.hasAccess(treatment.user.id, claims)) {
+            return await this.hospitalService.editHospitalTreatments(treatmentId, hospital);
+        }
     }
 
     @Delete('hospitals/:hospitalTreatmentId')
     @Roles(Role.Doctor)
     async deleteHospitalTreatments(
-        @Param('hospitalTreatmentId') treatmentId: string
+        @Param('hospitalTreatmentId') treatmentId: string,
+        @User() claims: Claims
     ): Promise<void>{
-        await this.hospitalService.deleteHospitalTreatments(treatmentId);
+        const treatment = await this.hospitalService.getUserId(treatmentId);
+        if (await this.doctorService.hasAccess(treatment.user.id, claims)) {
+            await this.hospitalService.deleteHospitalTreatments(treatmentId);
+        }
     }
-
-
-
  }
